requirements.txt
certifi==2018.11.29
chardet==3.0.4
darkskylib==0.3.91
dj-database-url==0.5.0
Django==2.1.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-heroku==0.3.1
future==0.17.1
googlemaps==3.0.2
gunicorn==19.9.0
idna==2.7
Pillow==5.3.0
psycopg2==2.7.6.1
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
whitenoise==4.1.2
